feat(snake): add hideProgressBar option to contribution counter
The hideProgressBar option is now configurable via the INPUT_HIDE_PROGRESS_BAR
environment variable. This allows users to control whether the progress bar
is displayed in the contribution counter animation.

The change adds a new property to the animationOptions.contributionCounter
object that reads from the environment variable and sets the hideProgressBar
flag accordingly.

Signed-off-by: diverger <diverger@live.cn>